redirect link registration, biggest challenge, day 1 & day 2

+++ application/controllers/HomeController.php
@@ -421,7 +421,11 @@ class HomeController extends CI_Controller {
 		$this->load->library('upload'); 
 		$this->upload->initialize($config);
 
-		unlink(PUBPATH . 'assets/user/uploads/'.$this->session->userdata('user_resume'));
+		if ($this->session->userdata('user_resume') != '') {
+
+			unlink(PUBPATH . 'assets/user/uploads/'.$this->session->userdata('user_resume'));
+		}
+		
 		if ( ! $this->upload->do_upload('uploaded_resume'))
 		{
 			$error = array('error' => $this->upload->display_errors());
